+++ src/loaders/settings/general.tsx
@@ -5,10 +5,11 @@ export default function GeneralSettingsLoader(){
           <div className="space-y-2">
                <SettingsItemLoader>
                     <SettingsOptionLoader optionType="selector" width={128}/>
+                    <SettingsOptionLoader optionType="brightness"/>
+                    <SettingsOptionLoader optionType="switch"/>
                </SettingsItemLoader>
                <SettingsItemLoader>
                     <SettingsOptionLoader optionType="selector" width={112}/>
-                    <SettingsOptionLoader optionType="switch"/>
                     <SettingsOptionLoader optionType="switch"/>
                </SettingsItemLoader>
           </div>